requirements.txt
# Python baseline: 3.10+

# Core trading
ccxt
backtrader

# Data processing
pandas
numpy

# Configuration
pyyaml
python-dotenv

# Logging and CLI
loguru
rich

# Visualization
matplotlib

# Development tools (optional)
pytest
pytest-asyncio
black
